ValueTypes: Use size_t not intptr_t for getRawBits
SimpleTy is a plain integer, and LLVMTy can be uniquely identified by
its address; the capability metadata is unnecessary. This avoids needing
to add overloads for (u)intcap_t to FoldingSet::AddInteger (which would
be needed because otherwise the existing integer primitive type
overloads are ambiguous due to multiple implicit conversions being
possible), which is all this function is ever used for.

+++ llvm/include/llvm/CodeGen/ValueTypes.h
@@ -424,11 +424,11 @@ namespace llvm {
     /// types are returned as Other, otherwise they are invalid.
     static EVT getEVT(Type *Ty, bool HandleUnknown = false);
 
-    intptr_t getRawBits() const {
+    size_t getRawBits() const {
       if (isSimple())
         return V.SimpleTy;
       else
-        return (intptr_t)(LLVMTy);
+        return (size_t)(intptr_t)(LLVMTy);
     }
 
     /// A meaningless but well-behaved order, useful for constructing
